Replace trainwreck shell script with php daemon.
This php daemon can run as non-root and create thumbnails
of video's, recode them with ffmpeg for the web and store all
results in the cache.

Use the systemd unit file to start it. Should need very little
configuration to actually get working.

Once started, it'll make sure all video's are recorded before
using inotify to screen for video uploads, so, on first startup
it will run for a while.