Make thumbnails of videos work.
This is a bit of a kludge that needs more cleanup, but
if the cache holds an unsized version of a file, it is fed
instead of searching for the proper sized version of a file.

This makes it easy to request .thm files or recoded video
formats without passing more flags to image.php.